requirements.txt
asgiref==3.4.1
black==21.9b0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
Django==3.2.8
django-rest-knox==4.1.0
djangorestframework==3.12.4
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pycparser==2.20
pytz==2021.3
regex==2021.10.23
sqlparse==0.4.2
tomli==1.2.2
typing-extensions==3.10.0.2
